Work on cartridge UI (buttons / switch / LED)

+++ C64/Cartridges/Cartridge.cpp
@@ -284,7 +284,7 @@ Cartridge::loadFromBuffer(uint8_t **buffer)
     setRamCapacity(read32(buffer));
     readBlock(buffer, externalRam, ramCapacity);
     persistentRam = (bool)read8(buffer);
-    switchPos = read8(buffer);
+    switchPos = (int8_t)read8(buffer);
     
     readBlock(buffer, val, sizeof(val));
     cycle = read64(buffer);
@@ -314,7 +314,7 @@ Cartridge::saveToBuffer(uint8_t **buffer)
     write32(buffer, ramCapacity);
     writeBlock(buffer, externalRam, ramCapacity);
     write8(buffer, (uint8_t)persistentRam);
-    write8(buffer, switchPos);
+    write8(buffer, (uint8_t)switchPos);
 
     writeBlock(buffer, val, sizeof(val));
     write64(buffer, cycle);
